Fix duplicate operation id
