+++ collections/m2m gateway/tenants/certifiedAttributes/List tenant certified attributes.bru
