+++ collections/m2m gateway/tenants/declaredAttributes/List tenant declared attributes.bru
