+++ collections/m2m gateway/tenants/verifiedAttributes/List tenant verified attributes.bru
