+++ packages/m2m-gateway/src/services/tenantService.ts
@@ -110,7 +110,7 @@ export function tenantServiceBuilder(clients: PagoPAInteropBeClients) {
 
       return toM2MGatewayApiTenant(tenant);
     },
-    async getDeclaredAttributes(
+    async getTenantDeclaredAttributes(
       tenantId: TenantId,
       {
         delegationId,
@@ -148,7 +148,7 @@ export function tenantServiceBuilder(clients: PagoPAInteropBeClients) {
         },
       };
     },
-    async getCertifiedAttributes(
+    async getTenantCertifiedAttributes(
       tenantId: TenantId,
       { limit, offset }: m2mGatewayApi.GetTenantCertifiedAttributesQueryParams,
       { logger, headers }: WithLogger<M2MGatewayAppContext>
@@ -227,7 +227,7 @@ export function tenantServiceBuilder(clients: PagoPAInteropBeClients) {
 
       return toM2MGatewayApiTenantCertifiedAttribute(certifiedAttribute);
     },
-    async getVerifiedAttributes(
+    async getTenantVerifiedAttributes(
       tenantId: TenantId,
       { limit, offset }: m2mGatewayApi.GetTenantVerifiedAttributesQueryParams,
       { logger, headers }: WithLogger<M2MGatewayAppContext>
