+++ packages/m2m-gateway/test/api/tenants/getTenantCertifiedAttributes.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from "vitest";
 import {
   generateToken,
-  getMockedApiVerifiedTenantAttribute,
+  getMockedApiCertifiedTenantAttribute,
 } from "pagopa-interop-commons-test";
 import { AuthRole, authRole } from "pagopa-interop-commons";
 import request from "supertest";
@@ -9,23 +9,25 @@ import { m2mGatewayApi } from "pagopa-interop-api-clients";
 import { generateId } from "pagopa-interop-models";
 import { api, mockTenantService } from "../../vitest.api.setup.js";
 import { appBasePath } from "../../../src/config/appBasePath.js";
-import { toM2MGatewayApiTenantVerifiedAttribute } from "../../../src/api/tenantApiConverter.js";
+import { toM2MGatewayApiTenantCertifiedAttribute } from "../../../src/api/tenantApiConverter.js";
 
-describe("GET /tenants/:tenantId/verifiedAttributes route test", () => {
-  const mockQueryParams: m2mGatewayApi.GetTenantVerifiedAttributesQueryParams =
+describe("GET /tenants/:tenantId/certifiedAttributes route test", () => {
+  const mockQueryParams: m2mGatewayApi.GetTenantCertifiedAttributesQueryParams =
     {
       offset: 0,
       limit: 10,
     };
 
-  const mockTenantAttribute1 = getMockedApiVerifiedTenantAttribute();
+  const mockTenantAttribute1 = getMockedApiCertifiedTenantAttribute();
 
-  const mockTenantAttribute2 = getMockedApiVerifiedTenantAttribute();
+  const mockTenantAttribute2 = getMockedApiCertifiedTenantAttribute({
+    revoked: true,
+  });
 
-  const mockResponse: m2mGatewayApi.TenantVerifiedAttributes = {
+  const mockResponse: m2mGatewayApi.TenantCertifiedAttributes = {
     results: [
-      toM2MGatewayApiTenantVerifiedAttribute(mockTenantAttribute1),
-      toM2MGatewayApiTenantVerifiedAttribute(mockTenantAttribute2),
+      toM2MGatewayApiTenantCertifiedAttribute(mockTenantAttribute1),
+      toM2MGatewayApiTenantCertifiedAttribute(mockTenantAttribute2),
     ],
     pagination: {
       limit: 10,
@@ -36,10 +38,10 @@ describe("GET /tenants/:tenantId/verifiedAttributes route test", () => {
 
   const makeRequest = async (
     token: string,
-    query: m2mGatewayApi.GetTenantVerifiedAttributesQueryParams
+    query: m2mGatewayApi.GetTenantCertifiedAttributesQueryParams
   ) =>
     request(api)
-      .get(`${appBasePath}/tenants/${generateId()}/verifiedAttributes`)
+      .get(`${appBasePath}/tenants/${generateId()}/certifiedAttributes`)
       .query(query)
       .set("Authorization", `Bearer ${token}`);
 
@@ -51,7 +53,7 @@ describe("GET /tenants/:tenantId/verifiedAttributes route test", () => {
   it.each(authorizedRoles)(
     "Should return 200 and perform service calls for user with role %s",
     async (role) => {
-      mockTenantService.getVerifiedAttributes = vi
+      mockTenantService.getTenantCertifiedAttributes = vi
         .fn()
         .mockResolvedValue(mockResponse);
 
@@ -81,7 +83,7 @@ describe("GET /tenants/:tenantId/verifiedAttributes route test", () => {
     const token = generateToken(authRole.M2M_ADMIN_ROLE);
     const res = await makeRequest(
       token,
-      query as m2mGatewayApi.GetTenantVerifiedAttributesQueryParams
+      query as m2mGatewayApi.GetTenantCertifiedAttributesQueryParams
     );
 
     expect(res.status).toBe(400);
@@ -107,7 +109,9 @@ describe("GET /tenants/:tenantId/verifiedAttributes route test", () => {
   ])(
     "Should return 500 when API model parsing fails for response",
     async (resp) => {
-      mockTenantService.getVerifiedAttributes = vi.fn().mockResolvedValue(resp);
+      mockTenantService.getTenantCertifiedAttributes = vi
+        .fn()
+        .mockResolvedValue(resp);
       const token = generateToken(authRole.M2M_ADMIN_ROLE);
       const res = await makeRequest(token, mockQueryParams);
