+++ packages/m2m-gateway/test/api/tenants/getTenantDeclaredAttributes.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from "vitest";
 import {
   generateToken,
-  getMockedApiCertifiedTenantAttribute,
+  getMockedApiDeclaredTenantAttribute,
 } from "pagopa-interop-commons-test";
 import { AuthRole, authRole } from "pagopa-interop-commons";
 import request from "supertest";
@@ -9,25 +9,25 @@ import { m2mGatewayApi } from "pagopa-interop-api-clients";
 import { generateId } from "pagopa-interop-models";
 import { api, mockTenantService } from "../../vitest.api.setup.js";
 import { appBasePath } from "../../../src/config/appBasePath.js";
-import { toM2MGatewayApiTenantCertifiedAttribute } from "../../../src/api/tenantApiConverter.js";
+import { toM2MGatewayApiTenantDeclaredAttribute } from "../../../src/api/tenantApiConverter.js";
 
-describe("GET /tenants/:tenantId/certifiedAttributes route test", () => {
-  const mockQueryParams: m2mGatewayApi.GetTenantCertifiedAttributesQueryParams =
+describe("GET /tenants/:tenantId/declaredAttributes route test", () => {
+  const mockQueryParams: m2mGatewayApi.GetTenantDeclaredAttributesQueryParams =
     {
       offset: 0,
       limit: 10,
     };
 
-  const mockTenantAttribute1 = getMockedApiCertifiedTenantAttribute();
+  const mockTenantAttribute1 = getMockedApiDeclaredTenantAttribute();
 
-  const mockTenantAttribute2 = getMockedApiCertifiedTenantAttribute({
+  const mockTenantAttribute2 = getMockedApiDeclaredTenantAttribute({
     revoked: true,
   });
 
-  const mockResponse: m2mGatewayApi.TenantCertifiedAttributes = {
+  const mockResponse: m2mGatewayApi.TenantDeclaredAttributes = {
     results: [
-      toM2MGatewayApiTenantCertifiedAttribute(mockTenantAttribute1),
-      toM2MGatewayApiTenantCertifiedAttribute(mockTenantAttribute2),
+      toM2MGatewayApiTenantDeclaredAttribute(mockTenantAttribute1),
+      toM2MGatewayApiTenantDeclaredAttribute(mockTenantAttribute2),
     ],
     pagination: {
       limit: 10,
@@ -38,10 +38,10 @@ describe("GET /tenants/:tenantId/certifiedAttributes route test", () => {
 
   const makeRequest = async (
     token: string,
-    query: m2mGatewayApi.GetTenantCertifiedAttributesQueryParams
+    query: m2mGatewayApi.GetTenantDeclaredAttributesQueryParams
   ) =>
     request(api)
-      .get(`${appBasePath}/tenants/${generateId()}/certifiedAttributes`)
+      .get(`${appBasePath}/tenants/${generateId()}/declaredAttributes`)
       .query(query)
       .set("Authorization", `Bearer ${token}`);
 
@@ -53,7 +53,7 @@ describe("GET /tenants/:tenantId/certifiedAttributes route test", () => {
   it.each(authorizedRoles)(
     "Should return 200 and perform service calls for user with role %s",
     async (role) => {
-      mockTenantService.getCertifiedAttributes = vi
+      mockTenantService.getTenantDeclaredAttributes = vi
         .fn()
         .mockResolvedValue(mockResponse);
 
@@ -83,7 +83,7 @@ describe("GET /tenants/:tenantId/certifiedAttributes route test", () => {
     const token = generateToken(authRole.M2M_ADMIN_ROLE);
     const res = await makeRequest(
       token,
-      query as m2mGatewayApi.GetTenantCertifiedAttributesQueryParams
+      query as m2mGatewayApi.GetTenantDeclaredAttributesQueryParams
     );
 
     expect(res.status).toBe(400);
@@ -109,7 +109,7 @@ describe("GET /tenants/:tenantId/certifiedAttributes route test", () => {
   ])(
     "Should return 500 when API model parsing fails for response",
     async (resp) => {
-      mockTenantService.getCertifiedAttributes = vi
+      mockTenantService.getTenantDeclaredAttributes = vi
         .fn()
         .mockResolvedValue(resp);
       const token = generateToken(authRole.M2M_ADMIN_ROLE);
